feat: add an option to fix Gitako's shrink feature
The fix in this commit will always show navigation on 1280px or higher
resolution.

Whiel it's not a perfect solution, it resolves an issue that causes some
navigation items to be hidden and the three-dot menu to appear on the
right side. Will be removed once this is fixed in Gitako.