fix: resolve conflicts for text colors of buttons
I'm switching disabled state of buttons from foreground to gray colors
in order to avoid issues with contrast. In my case, I'm using very light
orange/tan foreground color and it looks a lot better this way; might
need to play around with tones of gray colors tho.

One of the manual overrides for buttons is causing conflicts, so I made
styles for documentation apply only on the docs sub-domain.